backend/requirements.txt
Django==3.2.25
djangorestframework==3.14.0
django-cors-headers==3.14.0
djangorestframework-simplejwt==5.3.1
django-filter==22.1
djongo==1.3.6
pymongo==3.12.3
dnspython==2.6.1
python-dotenv==1.0.1
Pillow==10.4.0
sqlparse==0.2.4
